Fix error when reading in save
I forgot that the 32 byte header is stripped off by reading it and
was undershooting the end of the save data by quite a bit.

There's also some test functionality in here so I can decipher the
structure of this save data.

+++ src/main/java/com/woernerj/dragonsdogma/util/CompressionUtils.java
@@ -39,8 +39,8 @@ public class CompressionUtils {
 		try (ByteArrayOutputStream out = new ByteArrayOutputStream(compressed.length)){
 			byte[] buffer = new byte[1024];
 			while (inflater.getRemaining() > 0) {
-				double perc = inflater.getBytesRead()/(double)compressed.length;
-				System.out.printf("%.3f\r",perc);
+				float perc = inflater.getBytesRead()/(float)compressed.length;
+				System.out.printf("%f\r",perc);
 				out.write(buffer, 0, inflater.inflate(buffer));
 			}
 			inflater.end();
